RED Test that a legal move is legal
We are aiming to write Noughts and Crosses using Keith Braithwaite's
'Test-driven development as if you meant it' ethos.

We are starting from the is_legal (to be) method, that checks the
validity of a proposed move.  Our first test is somewhat contrived due
to the restrictions of the process.

It seems more natural to approach TDDAIYMI in a bottom-up, as opposed to
to-down, manner.