REFACTOR Improve design by moving code
The Rules allow us to move code to improve its structure.  This is not
really refactoring in the strictest sense, as it changes the behaviour
of the function (whereas refactoring should preserve behaviour).
However, the tests all still pass and this has given us a way to improve
the design.