RED Check if played moves win
This is another small step towards checking for a win (and who won).
Alas it has to be a somewhat contrived failure.  We could've put a
broader test here that required filtering out a player's moves, but we
are trying to take small steps, as per the rules (the hope being they
can be composed usefully later).

We have a predefined list of winning moves, rather than
generating/testing them at runtime; this is the simplest solution (and
the winning moves are inherent in the game).