REFACTOR Extract is_win() function
This moves the definition of winning moves to the is_win() function.  We
felt it was acceptable to do the `sorted(player_moves)` in is_win()
because the first test had the list in ascending order anyway (we could
have refactored to add the `sorted(player_moves)` to the first test, and
then refactored again to remove duplication).